migrations/requirements.txt
psycopg2-binary>=2.9.9
toml>=0.10.2
